InternalAPI timeouts
